split sendMail tests into mocked and non-mocked
